Server/requirements.txt
astroid==2.3.3
Click==7.0
Flask==1.1.1
Flask-API==2.0
Flask-WTF==0.14.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
PyJWT==1.7.1
pylint==2.4.4
python-dotenv==0.10.5
six==1.14.0
typed-ast==1.4.1
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
